moving conductor out of utils
